requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
chromadb==0.4.24
sentence-transformers==2.7.0
pdfplumber==0.11.7
pytesseract==0.3.10
Pillow==9.5.0
tqdm==4.67.1
python-dotenv==1.0.1
nltk==3.8.1
numpy==1.26.4
torch==2.1.0+cpu
opencv-python==4.8.1.78
pdf2image==1.17.0
pypdfium2==4.30.0
protobuf==3.20.3
layoutparser==0.3.4
openai==1.30.1
pydantic==1.10.13
tabulate==0.9.0
qdrant-client==1.8.0
fastapi==0.116.1
uvicorn
pdfplumber
pytesseract
fitz  # usually comes with PyMuPDF, or install as pymupdf
nltk
layoutparser
sentence-transformers
chromadb
python-dotenv
tqdm
opencv-python
Pillow
